Split context menu action widget
