+++ src/gui_app/context_menu.rs
@@ -1,17 +1,8 @@
 use super::GuiMessage;
+mod action_button;
+
+use radiant::gui::types::Point;
 use radiant::prelude as ui;
-use radiant::{
-    gui::types::{Point, Rect, Rgba8},
-    layout::{LayoutOutput, Vector2},
-    runtime::{
-        PaintFillRect, PaintPrimitive, PaintStrokeRect, PaintText, PaintTextAlign, PaintTextRun,
-    },
-    theme::ThemeTokens,
-    widgets::{
-        FocusBehavior, PaintBounds, PointerButton, TextWrap, Widget, WidgetCommon, WidgetInput,
-        WidgetOutput, WidgetSizing,
-    },
-};
 use std::path::{Path, PathBuf};
 
 const CONTEXT_MENU_WIDTH: f32 = 210.0;
@@ -154,10 +145,7 @@ fn context_menu_height(menu: &BrowserContextMenu) -> f32 {
 }
 
 fn context_menu_action(label: impl Into<String>, message: GuiMessage) -> ui::View<GuiMessage> {
-    ui::custom_widget_mapped(
-        ContextMenuActionButton::new(label.into(), message),
-        |message: GuiMessage| message,
-    )
+    action_button::view(label, message)
 }
 
 fn overlay_gap() -> ui::View<GuiMessage> {
@@ -169,243 +157,4 @@ fn dismiss_area(key: &'static str) -> ui::View<GuiMessage> {
         .message(GuiMessage::CloseContextMenu)
         .key(key)
         .input_only()
-}
-
-#[derive(Clone, Debug)]
-struct ContextMenuActionButton {
-    common: WidgetCommon,
-    label: String,
-    message: GuiMessage,
-}
-
-impl ContextMenuActionButton {
-    fn new(label: String, message: GuiMessage) -> Self {
-        let mut common = WidgetCommon::new(0, WidgetSizing::fixed(Vector2::new(1.0, 28.0)));
-        common.focus = FocusBehavior::None;
-        common.paint.bounds = PaintBounds::ClipToRect;
-        common.paint.paints_focus = false;
-        common.paint.paints_state_layers = false;
-        Self {
-            common,
-            label,
-            message,
-        }
-    }
-}
-
-impl Widget for ContextMenuActionButton {
-    fn common(&self) -> &WidgetCommon {
-        &self.common
-    }
-
-    fn common_mut(&mut self) -> &mut WidgetCommon {
-        &mut self.common
-    }
-
-    fn handle_input(&mut self, bounds: Rect, input: WidgetInput) -> Option<WidgetOutput> {
-        match input {
-            WidgetInput::PointerMove { position } => {
-                self.common.state.hovered = bounds.contains(position);
-                None
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                self.common.state.hovered = true;
-                self.common.state.pressed = true;
-                None
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } => {
-                let activated = self.common.state.pressed && bounds.contains(position);
-                self.common.state.pressed = false;
-                self.common.state.hovered = bounds.contains(position);
-                activated.then(|| WidgetOutput::typed(self.message.clone()))
-            }
-            _ => {
-                if matches!(input, WidgetInput::PointerRelease { .. }) {
-                    self.common.state.pressed = false;
-                }
-                None
-            }
-        }
-    }
-
-    fn accepts_pointer_move(&self) -> bool {
-        true
-    }
-
-    fn append_paint(
-        &self,
-        primitives: &mut Vec<PaintPrimitive>,
-        bounds: Rect,
-        _layout: &LayoutOutput,
-        _theme: &ThemeTokens,
-    ) {
-        let hovered = self.common.state.hovered || self.common.state.pressed;
-        primitives.push(PaintPrimitive::FillRect(PaintFillRect {
-            widget_id: self.common.id,
-            rect: bounds,
-            color: context_menu_button_fill(),
-        }));
-        if hovered {
-            primitives.push(PaintPrimitive::FillRect(PaintFillRect {
-                widget_id: self.common.id,
-                rect: bounds,
-                color: context_menu_hover_fill(self.common.state.pressed),
-            }));
-        }
-        primitives.push(PaintPrimitive::StrokeRect(PaintStrokeRect {
-            widget_id: self.common.id,
-            rect: Rect::from_min_max(
-                Point::new(bounds.min.x + 0.5, bounds.min.y + 0.5),
-                Point::new(bounds.max.x - 0.5, bounds.max.y - 0.5),
-            ),
-            color: context_menu_button_border(),
-            width: 1.0,
-        }));
-
-        let text_rect = Rect::from_min_max(
-            Point::new(bounds.min.x + 8.0, bounds.min.y),
-            Point::new(bounds.max.x - 8.0, bounds.max.y),
-        );
-        let font_size = 11.0;
-        primitives.push(PaintPrimitive::Text(PaintTextRun {
-            widget_id: self.common.id,
-            text: PaintText::from(self.label.as_str()),
-            rect: text_rect,
-            font_size,
-            baseline: Some(((text_rect.height() - font_size) * 0.5 + font_size * 0.78).round()),
-            color: context_menu_text(),
-            align: PaintTextAlign::Left,
-            wrap: TextWrap::None,
-        }));
-    }
-}
-
-fn context_menu_button_fill() -> Rgba8 {
-    Rgba8 {
-        r: 30,
-        g: 31,
-        b: 31,
-        a: 235,
-    }
-}
-
-fn context_menu_button_border() -> Rgba8 {
-    Rgba8 {
-        r: 66,
-        g: 67,
-        b: 68,
-        a: 175,
-    }
-}
-
-fn context_menu_hover_fill(pressed: bool) -> Rgba8 {
-    Rgba8 {
-        r: 255,
-        g: 108,
-        b: 88,
-        a: if pressed { 170 } else { 155 },
-    }
-}
-
-fn context_menu_text() -> Rgba8 {
-    Rgba8 {
-        r: 218,
-        g: 219,
-        b: 219,
-        a: 238,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use radiant::widgets::PointerModifiers;
-
-    fn action_primitives(button: &ContextMenuActionButton) -> Vec<PaintPrimitive> {
-        let mut primitives = Vec::new();
-        button.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(180.0, 28.0)),
-            &LayoutOutput::default(),
-            &ThemeTokens::default(),
-        );
-        primitives
-    }
-
-    #[test]
-    fn action_button_uses_hover_fill_without_recoloring_text() {
-        let mut button =
-            ContextMenuActionButton::new(String::from("Copy Path"), GuiMessage::CopyContextPath);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(180.0, 28.0));
-
-        let default_primitives = action_primitives(&button);
-        assert!(
-            !default_primitives.iter().any(|primitive| matches!(
-                primitive,
-                PaintPrimitive::FillRect(fill) if fill.color == context_menu_hover_fill(false)
-            )),
-            "{default_primitives:?}"
-        );
-        assert!(default_primitives.iter().any(|primitive| matches!(
-            primitive,
-            PaintPrimitive::Text(text) if text.color == context_menu_text()
-        )));
-
-        button.handle_input(
-            bounds,
-            WidgetInput::PointerMove {
-                position: Point::new(20.0, 12.0),
-            },
-        );
-
-        let hover_primitives = action_primitives(&button);
-        assert!(hover_primitives.iter().any(|primitive| matches!(
-            primitive,
-            PaintPrimitive::FillRect(fill)
-                if fill.rect == bounds && fill.color == context_menu_hover_fill(false)
-        )));
-        assert!(hover_primitives.iter().any(|primitive| matches!(
-            primitive,
-            PaintPrimitive::Text(text) if text.color == context_menu_text()
-        )));
-    }
-
-    #[test]
-    fn action_button_emits_configured_message_on_click() {
-        let mut button =
-            ContextMenuActionButton::new(String::from("Copy Path"), GuiMessage::CopyContextPath);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(180.0, 28.0));
-
-        button.handle_input(
-            bounds,
-            WidgetInput::PointerPress {
-                position: Point::new(20.0, 12.0),
-                button: PointerButton::Primary,
-                modifiers: PointerModifiers::default(),
-            },
-        );
-        let output = button
-            .handle_input(
-                bounds,
-                WidgetInput::PointerRelease {
-                    position: Point::new(20.0, 12.0),
-                    button: PointerButton::Primary,
-                    modifiers: PointerModifiers::default(),
-                },
-            )
-            .expect("click should emit action message");
-
-        assert_eq!(
-            output.typed_ref::<GuiMessage>(),
-            Some(&GuiMessage::CopyContextPath)
-        );
-    }
 }
